Split randomness into a package
